Move getCmdStatus to job command; modify stress bench stop time
